% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aquacrop.R
\name{.install_templates}
\alias{.install_templates}
\title{Install simulation template files}
\usage{
.install_templates(path, overwrite)
}
\description{
Install simulation template files
}
\keyword{internal}
