% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.normalize_param_bounds}
\alias{.normalize_param_bounds}
\title{Normalise parameter bounds to a data.frame of name/min/max}
\usage{
.normalize_param_bounds(params)
}
\description{
Normalise parameter bounds to a data.frame of name/min/max
}
\keyword{internal}
