% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity.R
\name{ece_to_salinity}
\alias{ece_to_salinity}
\title{Get Salinity Description from ECe Value}
\usage{
ece_to_salinity(ece)
}
\arguments{
\item{ece}{Numeric. Electrical conductivity in dS/m}
}
\value{
Character string describing the salinity level
}
\description{
Convert electrical conductivity (ECe) values to qualitative salinity descriptions.
}
\examples{
ece_to_salinity(0) # "Non-saline"
ece_to_salinity(3.5) # "Slightly saline"
ece_to_salinity(10) # "Strongly saline"

}
