% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paths.R
\name{get_aquacrop_version}
\alias{get_aquacrop_version}
\title{Get AquaCrop Version}
\usage{
get_aquacrop_version()
}
\value{
Character string with version information.
}
\description{
Returns the version of the bundled AquaCrop executable.
}
\examples{
\dontrun{
version <- get_aquacrop_version()
print(version)
}

}
