% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_cli}
\alias{is_cli}
\title{Check if file is a valid AquaCrop climate file}
\usage{
is_cli(file)
}
\arguments{
\item{file}{Path to the file to check}
}
\value{
Logical indicating if file is a valid .CLI file
}
\description{
Check if file is a valid AquaCrop climate file
}
