% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_eto}
\alias{is_eto}
\title{Check if file is a valid AquaCrop reference evapotranspiration file}
\usage{
is_eto(file)
}
\arguments{
\item{file}{Path to the file to check}
}
\value{
Logical indicating if file is a valid .ETo file
}
\description{
Check if file is a valid AquaCrop reference evapotranspiration file
}
