% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check if a Year is a Leap Year}
\usage{
is_leap_year(x)
}
\arguments{
\item{x}{A numeric scalar representing a year}
}
\value{
A logical scalar (TRUE if leap year, FALSE otherwise)
}
\description{
Check if a Year is a Leap Year
}
\examples{
is_leap_year(2020)
is_leap_year(2021)

}
\seealso{
Other utility functions: 
\code{\link{calculate_plant_density}()},
\code{\link{path_to_file}()},
\code{\link{round_to}()},
\code{\link{salinity_to_ece}()},
\code{\link{to_aquacrop_day}()}
}
\concept{utility functions}
