% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_tnx}
\alias{is_tnx}
\title{Check if file is a valid AquaCrop temperature file}
\usage{
is_tnx(file)
}
\arguments{
\item{file}{Path to the file to check}
}
\value{
Logical indicating if file is a valid .Tnx file
}
\description{
Check if file is a valid AquaCrop temperature file
}
