% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_defaults}
\alias{parameter_defaults}
\title{Default values for every parameter}
\usage{
parameter_defaults(parameters)
}
\arguments{
\item{parameters}{An \code{aquacropr_parameters} object.}
}
\value{
Named numeric vector.
}
\description{
Default values for every parameter
}
