% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_names}
\alias{parameter_names}
\title{Names of every free parameter}
\usage{
parameter_names(parameters)
}
\arguments{
\item{parameters}{An \code{aquacropr_parameters} object.}
}
\value{
Character vector.
}
\description{
Names of every free parameter
}
