% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{predict.aquacropr_metamodel}
\alias{predict.aquacropr_metamodel}
\title{Predict from a fitted AquaCrop metamodel}
\usage{
\method{predict}{aquacropr_metamodel}(object, newdata, type = "UK", ...)
}
\arguments{
\item{object}{An \code{aquacropr_metamodel}.}

\item{newdata}{A data.frame with \code{object$param_names}.}

\item{type}{\code{"UK"} or \code{"SK"}.}

\item{...}{Forwarded to DiceKriging's \code{predict} method.}
}
\description{
Predict from a fitted AquaCrop metamodel
}
