% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read parameters from a YAML file}
\usage{
read_parameters(path)
}
\arguments{
\item{path}{Character scalar path.}
}
\value{
An \code{aquacropr_parameters} object.
}
\description{
The file should have a top-level \verb{parameters:} list. Each entry needs
\code{name}, \code{default}, \code{min}, \code{max}, \code{from_file}, and \code{to_file}.
}
