% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{render_one_template}
\alias{render_one_template}
\title{Render a single text template given raw substitutions}
\usage{
render_one_template(subs, from_file, to_file)
}
\arguments{
\item{subs}{Named list/vector of placeholder -> replacement text.}

\item{from_file}{Template path.}

\item{to_file}{Destination path.}
}
\value{
Invisibly, the rendered text.
}
\description{
Render a single text template given raw substitutions
}
\keyword{internal}
