% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_source.R
\name{build_source}
\alias{build_source}
\title{Build AquaCrop from Source}
\usage{
build_source(source_dir, compiler = "gfortran", target = "bin", verbose = TRUE)
}
\arguments{
\item{source_dir}{Character. Path to AquaCrop source directory.}

\item{compiler}{Character. Fortran compiler. Default: "gfortran".}

\item{target}{Character. Build target: "bin" (default), "lib", or "all".}

\item{verbose}{Logical. If TRUE will print building informations.}
}
\value{
Path to compiled executable.
}
\description{
Compiles AquaCrop executable from source code using make and Fortran compiler.
}
\examples{
\dontrun{
exe_path <- build_source("AquaCrop")
exe_path <- build_source("AquaCrop", compiler = "ifort")
}

}
