% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aquacrop.R
\name{.create_readme}
\alias{.create_readme}
\title{Helper function to create README}
\usage{
.create_readme(path, dirs, version = NULL)
}
\description{
Helper function to create README
}
\keyword{internal}
