% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binaries.R
\name{install_binaries}
\alias{install_binaries}
\title{Install AquaCrop binaries}
\usage{
install_binaries(version = NULL, os = NULL, path, force = FALSE)
}
\arguments{
\item{version}{Version (NULL = latest, e.g. "7.1"). Minimum: 7.0}

\item{os}{OS ("windows", "linux", "macos", NULL = auto-detect)}

\item{path}{Install path (single directory, no subdirs)}

\item{force}{Force reinstall (default: FALSE)}
}
\value{
Path to executable
}
\description{
Install AquaCrop binaries
}
