% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check if a Year is a Leap Year}
\usage{
is_leap_year(x)
}
\arguments{
\item{x}{A numeric scalar representing a year}
}
\value{
A logical scalar (TRUE if leap year, FALSE otherwise)
}
\description{
Check if a Year is a Leap Year
}
\examples{
is_leap_year(2020)
is_leap_year(2021)

}
