% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_plu}
\alias{is_plu}
\title{Check if file is a valid AquaCrop rainfall file}
\usage{
is_plu(file)
}
\arguments{
\item{file}{Path to the file to check}
}
\value{
Logical indicating if file is a valid .PLU file
}
\description{
Check if file is a valid AquaCrop rainfall file
}
